requirements.txt
python-dotenv==1.0.0
joblib==1.3.2
kafka-python==2.0.2
matplotlib==3.7.2
numpy==1.25.2
pandas==2.0.3
psycopg2==2.9.7
requests==2.27.1
scikit-learn==1.0.2
seaborn==0.12.2
uszipcode==1.0.1
